+++ backend/src/users/dto/create-user.dto.ts
@@ -1,1 +1,8 @@
-export class CreateUserDto {}
+import { PartialType } from '@nestjs/mapped-types';
+
+export class CreateUserDto {
+  email!: string;
+  password!: string;
+  firstName!: string;
+  lastName!: string;
+}